passando as informaçoes da home para home da pasta
